Finally rename Git_Hook to GitHook
This will break any conf files that directly reference class paths
