+++ src/Bart/GitHook/GitHookException.php
@@ -1,5 +1,5 @@
 <?php
-namespace Bart\Git_Hook;
+namespace Bart\GitHook;
 
 /**
  * Exception with git hooks
