+++ test/Bart/GitHook/Gerrit_Approved_Test.php
@@ -1,5 +1,5 @@
 <?php
-namespace Bart\Git_Hook;
+namespace Bart\GitHook;
 
 class Gerrit_Approved_Test extends TestBase
 {
